# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * AlWatan Ready-Mix Company public website and admin CMS API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export interface Product {
  id: number;
  nameAr: string;
  nameEn: string;
  slug: string;
  descriptionAr: string;
  descriptionEn: string;
  specsAr: string;
  specsEn: string;
  imageUrl: string;
  compressiveStrength: string;
}

export interface ProductInput {
  nameAr: string;
  nameEn: string;
  slug: string;
  descriptionAr: string;
  descriptionEn: string;
  specsAr: string;
  specsEn: string;
  imageUrl: string;
  compressiveStrength: string;
}

export type ProductUpdate = ProductInput;

export interface Branch {
  id: number;
  nameAr: string;
  nameEn: string;
  addressAr: string;
  addressEn: string;
  city: string;
  lat: number;
  lng: number;
  phone: string;
  workingHours: string;
  orderIndex: number;
}

export interface Certificate {
  id: number;
  titleAr: string;
  titleEn: string;
  descriptionAr: string;
  descriptionEn: string;
  imageUrl: string;
  issuedBy: string;
  issuedDate: string;
}

export type BlogPostStatus = typeof BlogPostStatus[keyof typeof BlogPostStatus];


export const BlogPostStatus = {
  draft: 'draft',
  published: 'published',
} as const;

export interface BlogPost {
  id: number;
  titleAr: string;
  titleEn: string;
  excerptAr: string;
  excerptEn: string;
  contentAr: string;
  contentEn: string;
  slug: string;
  coverImage: string;
  status: BlogPostStatus;
  publishedAt: string;
}

export type BlogPostInputStatus = typeof BlogPostInputStatus[keyof typeof BlogPostInputStatus];


export const BlogPostInputStatus = {
  draft: 'draft',
  published: 'published',
} as const;

export interface BlogPostInput {
  titleAr: string;
  titleEn: string;
  excerptAr: string;
  excerptEn: string;
  contentAr: string;
  contentEn: string;
  slug: string;
  coverImage: string;
  status: BlogPostInputStatus;
}

export type QuoteRequestStatus = typeof QuoteRequestStatus[keyof typeof QuoteRequestStatus];


export const QuoteRequestStatus = {
  new: 'new',
  contacted: 'contacted',
  closed: 'closed',
} as const;

export interface QuoteRequest {
  id: number;
  productId: number;
  productName: string;
  quantityEstimate: string;
  projectLocation: string;
  fullName: string;
  phone: string;
  email: string;
  notes: string;
  status: QuoteRequestStatus;
  createdAt: string;
}

export interface QuoteInput {
  productId: number;
  quantityEstimate: string;
  projectLocation: string;
  fullName: string;
  phone: string;
  email: string;
  notes?: string;
  website?: string;
}

export type QuoteStatusUpdateStatus = typeof QuoteStatusUpdateStatus[keyof typeof QuoteStatusUpdateStatus];


export const QuoteStatusUpdateStatus = {
  new: 'new',
  contacted: 'contacted',
  closed: 'closed',
} as const;

export interface QuoteStatusUpdate {
  status: QuoteStatusUpdateStatus;
}

export type PageContentContentType = typeof PageContentContentType[keyof typeof PageContentContentType];


export const PageContentContentType = {
  text: 'text',
  richtext: 'richtext',
  url: 'url',
} as const;

export interface PageContent {
  id: number;
  pageSlug: string;
  sectionKey: string;
  fieldKey: string;
  valueAr: string;
  valueEn: string;
  contentType: PageContentContentType;
  updatedAt: string;
}

export interface PageContentUpdate {
  valueAr: string;
  valueEn: string;
}

export interface PublicSettings {
  whatsapp: string;
  quoteEmail: string;
  phone: string;
  email: string;
  instagramUrl: string;
  linkedinUrl: string;
}

export interface AdminSummary {
  newQuotes: number;
  totalQuotes: number;
  publishedPosts: number;
  products: number;
  branches: number;
  recentQuotes: QuoteRequest[];
}

export interface HomePayload {
  products: Product[];
  branches: Branch[];
  certificates: Certificate[];
  blogPosts: BlogPost[];
  content: PageContent[];
  settings: PublicSettings;
}

/**
 * Resource not found
 */
export type NotFoundResponse = Error;