# lib/api-zod/src/generated/types/branch.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * AlWatan Ready-Mix Company public website and admin CMS API
 * OpenAPI spec version: 0.1.0
 */

export interface Branch {
  id: number;
  nameAr: string;
  nameEn: string;
  addressAr: string;
  addressEn: string;
  city: string;
  lat: number;
  lng: number;
  phone: string;
  workingHours: string;
  orderIndex: number;
}